Add hack for searchAttributes list
I want to add an entry like the following to .env:

  LDAPAUTH_SEARCHATTRIBUTES=["mail", "displayName", "memberOf"]

such that the config.json file will have a corresponding entry:

  "searchAttributes": ["mail", "displayName", "memberOf"]

But I can't seem to make it work. Maybe the awk commands in setconfig
that handle LDAPAUTH_ aren't set up to handle list entries yet. In any
event, I haven't yet taken the time to understand the details in the
chunk of setconfig that handles the LDAPAUTH_ variables.

In order to show a proof-of-principle for using authentication groups,
for the moment, adding this bad hack.

While we're here making hacks, hard-coding debug = true, because I
cannot seem to get that to work from .env either.