Check for authorized_group in /verify
This is similar to what was done for the /authorized endpoint in
2161769.

Add an option to send 'authorized_group' in the POST request body to
/verify endpoint. If the authorized_group is in the decoded token (and
the token is still valid), then the user is authenticated. If the
authorized_group is NOT in the decoded token, a 401 is returned:

{
  "error": "User not in authorized_group"
}

If no 'authorized_group' is in the initial POST request to /verify,
the behavior is unchanged from before.