Change payload key name in attempt at more clarity
Because of changes in a21b9ed, we are returning NOT the original list
of authorized_groups from the /authenticate request, but instead the
intersection of the user's groups and the original list of
authorized_groups from the /authenticate request. To try and make things
a bit more clear for developers, changing this key to
"users_authorized_groups" in the JWT payload.

Note that this key name change has no effect for users of the LDAP-JWT
service; this is just to try and make things more clear for developers.