Handle searchAttributes as 'list' in setconfig
Part of the difficulty in setting LDAPAUTH_ options that are to be lists
in config.json is that Docker does not handle array environment variables
(e.g. myarray=("a","b","c")) the same way bash does.

To try an help users of this service, provide an example of how to set
searchAttributes, and provide a sensible default value (as has been done
for searchFilter).