Change name of example config file
Change the file name to make it more clear this is an example. Also
removed the default value for this file in the test script, and added
error trapping.

This is all to help other developers notice the requirement of the
configuration file of test usernames, passwords, and authorized groups.