Add python testing virtual environment info
The unit tests for this project are written in python (I know..they
should be in node since the rest of the project is in node). Because I
had some python package version issues when running them today, I added
info to create a virtual environment for running the tests.