Add option for ci build
For running a GitHub Actions Worflow for CI testing, we don't need to
run a docker container. We can just use a runner with node installed.

With this in mind, giving setconfig the option to exit after writing the
config file in a CI build, and for index.js to start the mock ldap
server in a CI build.