Add default SEARCHFILTER in setconfig
We cannot add a reasonable default such as (CN={{username}}) in the
docker-compose.yml file because there is no way to escape the curly
braces.

If we DON'T have a default, then requiring LDAP_SEARCHFILTER is a
breaking change to 1.3.0.

So, to make the search filter configurable, w/o having a breaking
change, adding a default value in setconfig.